Slim TimePoint to three methods; narrow Timestamp to u64 nanoseconds
Per D11/D12 of the v2 decision record (stage 2).

D11: `TimePoint` demanded four methods, two of which nothing called.
`checked_add` had zero call sites in the crate, and `as_seconds` existed
only to feed the default `as_seconds_lossy` — whose own docs told
implementors to override it, so the integrator writing a clock for an MCU
implemented a double-precision divide that only ran if they forgot to.
Both are gone. `as_seconds_lossy` becomes required, which states in the
type system what the crate already depends on: error formatting cannot
fail, so a conversion error can never mask the error being reported.
`Debug` joins `Copy + Ord` as a supertrait — without it a clock type
lacking its own derive made `Transform<YourClock>` unprintable exactly
where a diagnosis is needed.

D12: `Timestamp` stores u64 nanoseconds instead of u128. u64 spans ~584
years — mid-2554 from the Unix epoch, past the service life of anything
this crate positions — while halving per-sample stamp storage and
removing multi-word arithmetic from the 32-bit MCU targets the charter
names first-class. `from_nanos`/`as_nanos` are public, so the width was a
now-or-never decision. The narrowing pays for itself inside the
arithmetic: every difference of two u64 nanosecond stamps is a valid
`Duration`, so `Sub` collapses to a single checked subtraction with no
seconds/nanos split, no truncating casts, and no unreachable overflow
arm. It also creates one new failure: a system clock past 2554 no longer
fits, which `try_now` reports as `DurationOverflow` and `now()`
documents — a panic message naming only the pre-epoch cause would have
misdiagnosed the other half, which is the failure mode this crate exists
to prevent, one level up in the diagnostic path.

The wire format is unaffected. Postcard's LEB128 varint is identical for
every value that fits both widths, and the golden-byte tests pass
byte-for-byte unchanged; MessagePack stops emitting the 16-byte blob a
u128 forced, so foreign-language consumers now read a plain integer.

Tests: the arithmetic that changed is pinned rather than assumed — the
full-range subtraction, the top of the range as ordinary data, and a
`Duration` too wide for the timestamp type rejected in both directions.
This stage also discharges the audit's proptest gap while it is here:
generated timestamps were capped at 1e15 ns, below the 2^53 `as_seconds`
cliff and 1000x below wall clock, so the magnitudes where precision bugs
actually live were untested. The strategy now draws from four bands —
low, straddling 2^53, 2020s wall-clock, and the top of the u64 range.

+++ benches/benchmarks.rs
@@ -8,14 +8,14 @@ use transforms::{
 };
 
 /// Base timestamp for dynamic samples.
-const BASE_NANOS: u128 = 1_000_000_000;
+const BASE_NANOS: u64 = 1_000_000_000;
 /// Nanoseconds between consecutive samples in the prepared registries.
-const SAMPLE_INTERVAL_NANOS: u128 = 1_000_000;
+const SAMPLE_INTERVAL_NANOS: u64 = 1_000_000;
 
 fn transform_at(
     parent: &str,
     child: &str,
-    nanos: u128,
+    nanos: u64,
 ) -> Transform {
     Transform {
         translation: Vector3::new(1.0, 0.0, 0.0),
@@ -29,7 +29,7 @@ fn transform_at(
 /// A registry pre-warmed with `samples` dynamic transforms between "a" and
 /// "b", spaced `SAMPLE_INTERVAL_NANOS` apart starting at `BASE_NANOS`.
 /// Returns the registry and the first free timestamp after the samples.
-fn prewarmed_registry(samples: u128) -> (Registry, u128) {
+fn prewarmed_registry(samples: u64) -> (Registry, u64) {
     let mut registry = Registry::new();
     let mut nanos = BASE_NANOS;
     for _ in 0..samples {
@@ -160,7 +160,7 @@ fn benchmark_get_transform_interpolated(c: &mut Criterion) {
 /// The three dynamic edges hold `samples` samples each; the three mounts
 /// are static. Deep synthetic chains overstate real-world lookup costs;
 /// this is the shape most robots actually query.
-fn robot_tree(samples: u128) -> Registry {
+fn robot_tree(samples: u64) -> Registry {
     let mut registry = Registry::new();
 
     registry
